Refactor API handlers and update OpenAPI paths
This commit reorganizes the API handlers by introducing modular submodules for `bundle`, `checkpoint`, `dashboard`, `health`, and `meta`. The OpenAPI paths in the `ApiAgentsQuery` struct are updated to reflect the new handler structure, enhancing the clarity and maintainability of the API. This refactor improves the overall organization of the codebase, facilitating easier navigation and future development.
